[Dembo] Refactor extract standard message to own file
