+++ plugins/slack-notification-plugin/slack/message/message.go
@@ -1,25 +1,5 @@
 package message
 
-import "encoding/json"
-
 type Message interface {
 	JSON() (string, error)
-}
-
-type standardMessage struct {
-	Text string `json:"text"`
-}
-
-func (messageObject *standardMessage) JSON() (string, error) {
-	byteMessage, err := json.Marshal(messageObject)
-	if err != nil {
-		return "", err
-	}
-	return string(byteMessage), nil
-}
-
-func NewStandardMessage(text string) Message {
-	return &standardMessage{
-		Text: text,
-	}
 }
